+++ src/components/Logo/Logo.tsx
@@ -2,70 +2,51 @@ import * as React from "react";
 import { Icon as ChakraIcon, Box, useStyleConfig } from "@chakra-ui/react";
 
 import generateUUID from "../../helpers/generateUUID";
-import {
-  IconAlign,
-  IconColors,
-  IconNames,
-  IconRotationTypes,
-  IconSizes,
-  IconTypes,
-} from "./IconTypes";
-import iconSvgs from "./IconSvgs";
+import { LogoColors, LogoNames, LogoSizes } from "./LogoTypes";
+import logoSvgs from "./LogoSvgs";
 
-export interface IconProps {
+export interface LogoProps {
   /** Optionally pass in additional Chakra-based styles. */
   additionalStyles?: { [key: string]: any };
-  /** Aligns the icon. */
-  align?: IconAlign;
   /** Optional className that will be added to the parent element */
   className?: string;
-  /** Overrides default icon color (black). */
-  color?: IconColors;
-  /** Icons designated as decorative will be ignored by screenreaders. True
+  /** Overrides default logo color (black). */
+  color?: LogoColors;
+  /** Logos designated as decorative will be ignored by screenreaders. False
    * by default. */
   decorative?: boolean;
-  /** Rotates the icon clockwise in increments of 90deg */
-  iconRotation?: IconRotationTypes;
   /** ID that other components can cross reference for accessibility purposes */
   id?: string;
-  /** The name of the icon you want to use. */
-  name?: IconNames;
-  /** Sets the icon size. */
-  size?: IconSizes;
+  /** The name of the logo you want to use. */
+  name?: LogoNames;
+  /** Sets the logo size. */
+  size?: LogoSizes;
   /** For accessibility purposes, the text passed in the `title` prop gets
    * rendered in a `title` element in the SVG. This descriptive text is not
    * visible but is needed for screenreaders to describe the graphic. */
   title?: string;
-  /** FOR INTERNAL DS USE ONLY: the icon variant to display. */
-  type?: IconTypes;
 }
 
 /**
- * Renders SVG-based icons.
+ * Renders SVG-based logos that commonly used by the New York Public Library.  Plase note that the overall color of a `color` logo can not be changed.
  */
-export default function Icon(props: React.PropsWithChildren<IconProps>) {
+export default function Logo(props: React.PropsWithChildren<LogoProps>) {
   const {
     additionalStyles = {},
-    align = IconAlign.None,
     children,
     className,
-    color = IconColors.UiBlack,
-    decorative = true,
-    iconRotation = IconRotationTypes.Rotate0,
+    color = LogoColors.UiBlack,
+    decorative = false,
     id = generateUUID(),
     name,
-    size = IconSizes.Medium,
-    title = `${name} icon`,
-    type = IconTypes.Default,
+    size = LogoSizes.Medium,
+    title = `${name} logo`,
   } = props;
-  const styles = useStyleConfig("Icon", {
-    align,
+  const styles = useStyleConfig("Logo", {
     color,
-    iconRotation,
     size,
-    variant: type,
   });
-  const iconProps = {
+  const logoProps = {
     "aria-hidden": decorative,
     className,
     id,
@@ -77,42 +58,42 @@ export default function Icon(props: React.PropsWithChildren<IconProps>) {
   // Component prop validation
   if (name && children) {
     console.warn(
-      "Icon accepts either a `name` prop or an `svg` element child, not both."
+      "Logo accepts either a `name` prop or an `svg` element child. It can not accept both."
     );
     return null;
   } else if (!name && !children) {
     console.warn(
-      "Pass an icon `name` prop or an SVG child to `Icon` to ensure an icon appears."
+      "Pass a logo `name` prop or an SVG child to `Logo` to ensure a logo appears."
     );
     return null;
   }
 
-  // The user wants to render an existing icon. Load the icon and render it
+  // The user wants to render an existing logo. Load the logo and render it
   // as a component through Chakra's Icon component. Otherwise, we're going to
   // render the SVG child with NYPL-theme styling.
   if (name) {
-    const SvgComponent: any = iconSvgs[name];
+    const SvgComponent: any = logoSvgs[name];
     return (
       <ChakraIcon
         as={SvgComponent}
-        {...iconProps}
+        {...logoProps}
         __css={{ ...styles, ...additionalStyles }}
       />
     );
   }
 
   // If no `name` prop was passed, we expect a child SVG element to be passed.
-  // Apply icon props to the SVG child.
+  // Apply logo props to the SVG child.
   if (
     (children as JSX.Element).type === "svg" ||
     (children as JSX.Element).props.type === "svg" ||
     (children as JSX.Element).props.mdxType === "svg"
   ) {
     childSVG = React.cloneElement(children as JSX.Element, {
-      ...iconProps,
+      ...logoProps,
     });
   } else {
-    console.warn("You must pass an `svg` element to the `Icon` component.");
+    console.warn("You must pass an `svg` element to the `Logo` component.");
   }
 
   return <Box __css={styles}>{childSVG}</Box>;