+++ src/components/Logo/LogoTypes.tsx
@@ -1,0 +1,47 @@
+export enum LogoColors {
+  UiBlack = "ui.black",
+  UiWhite = "ui.white",
+  BrandPrimary = "brand.primary",
+  BrandSecondary = "brand.secondary",
+  SectionWhatsOnPrimary = "section.whats-on.primary",
+  SectionWhatsOnSecondary = "section.whats-on.secondary",
+  SectionBooksAndMorePrimary = "section.books-and-more.primary",
+  SectionBooksAndMoreSecondary = "section.books-and-more.secondary",
+  SectionResearchPrimary = "section.research.primary",
+  SectionResearchSecondary = "section.research.secondary",
+  SectionLocationsPrimary = "section.locations.primary",
+  SectionLocationsSecondary = "section.locations.secondary",
+  SectionBlogsPrimary = "section.blogs.primary",
+  SectionBlogsSecondary = "section.blogs.secondary",
+  SectionResearchLibraryLpa = "section.research-library.lpa",
+  SectionResearchLibrarySchomburg = "section.research-library.schomburg",
+  SectionResearchLibrarySchwartzman = "section.research-library.schwartzman",
+}
+
+export enum LogoSizes {
+  ExtraExtraSmall = "xxsmall",
+  ExtraSmall = "xsmall",
+  Small = "small",
+  Medium = "medium",
+  Large = "large",
+  Default = "default",
+}
+
+export enum LogoNames {
+  BrooklynPublicLibrary = "logo_bpl",
+  LPA = "logo_lpa",
+  LPAColor = "logo_lpa_color",
+  NYPL = "logo_nypl",
+  NYPLLion = "logo_nypl_lion",
+  NYPLLionInverted = "logo_nypl_lion_inverted",
+  NYPLNoLion = "logo_nypl_no_lion",
+  QueensPublicLibrary = "logo_qpl",
+  QueensPublicLibraryAlt = "logo_qpl_horizontal",
+  Schomburg = "logo_schomburg",
+  SchomburgColor = "logo_schomburg_color",
+  SchomburgCircle = "logo_schomburg_circle",
+  SchomburgCircleColor = "logo_schomburg_circle_color",
+  SimplyE = "logo_simplye",
+  SimplyEColor = "logo_simplye_color",
+  SNFL = "logo_snfl",
+}